Added EventMaker.py and created a thread and a queue for it.
Now HwReader uses a queue to move data from inputPoller to EventMaker,
which both run in seperate threads from the main HwReader.

Queue is configured to prioritize Encoder inputs and to update the queue
with a fresh Encoder value when available.

Now the inputPoller is not held back by the speed of EventMaker.
(Note: Currently EventMaker is artificially throttled to simulate high load.)